modifs fin  On branch sophie  Changes to be committed: modified:   src/main/java/reservation/dao/DaoAeroport.java 	modified:   src/main/java/reservation/dao/DaoAeroportJpaImpl.java 	modified:   src/main/java/reservation/dao/DaoPassagerJpaImpl.java 	modified:   src/main/java/reservation/dao/DaoReservationJpaImpl.java 	modified:   src/main/java/reservation/dao/DaoVol.java 	modified:   src/main/java/reservation/dao/DaoVolJpaImpl.java 	modified:   src/main/java/reservation/model/Reservation.java 	modified:   src/test/java/reservation/AppTest.java

+++ src/main/java/reservation/dao/DaoAeroport.java
@@ -4,14 +4,14 @@ import java.util.List;
 
 import reservation.model.Aeroport;
 
-public interface DaoAeroport extends DaoGeneric<Aeroport, Integer> {
-	
-	public Aeroport findByKeyWithVille(Integer key);
-	
+public interface DaoAeroport extends DaoGeneric<Aeroport, Long> {
+
+	public Aeroport findByKeyWithVille(Long key);
+
 	public List<Aeroport> findAllWithVille();
-	
-	public Aeroport findByKeyWithEscale(Integer key);
-	
+
+	public Aeroport findByKeyWithEscale(Long key);
+
 	public List<Aeroport> findAllWithEscale();
 
 }
